+++ CHANGELOG.md
@@ -6,6 +6,7 @@ New features:
 * Add `Java.add_to_classpath` method to add jar paths at runtime (#2693, @bjfish).
 * Add support for Ruby 3.1's Hash shorthand/punning syntax (@nirvdrum).
 * Add support for Ruby 3.1's anonymous block forwarding syntax (@nirvdrum).
+* Added the following keyword arguments to `Polyglot::InnerContext.new`: `languages, language_options, inherit_all_access, code_sharing` (@eregon).
 
 Bug fixes: